* GameValue의 이름을 Resource로 바꾸고, 게임에서 사용될 자원들을 관리하도록 설정 * SoundManager 클래스가 사운드 재생에 관한 모든 것을 총괄하도록 수정
* SelectSong 씬에 진입할 때마다 매번 데이터를 파싱하여 Song 객체를 만들고, 지우는 과정 대신, Resource에서 Song 벡터를 가지도록 수정

* Song은 더이상 Sound객체를 갖지 않도록 수정

+++ Game/Manager/ConstValue.h
@@ -19,7 +19,7 @@ namespace ConstValue
 	const int MaxSync = 10000;
 	const int MinSync = -10000;
 
-	const float LoopLength = 850000.0f / 44000.0f;
+	const UINT32 LoopLength = static_cast<UINT32>(850000.0f / 44000.0f);
 
 	const float MaxSpeed = 20.0f;
 	const float DefSpeed = 6.0f;
@@ -46,10 +46,6 @@ namespace ConstValue
 	Difficulty operator++(Difficulty& d);
 	Difficulty operator--(Difficulty& d);
 	std::string const& ToString(Difficulty& d);
-
-
-
-	
 }
 
 namespace Keycode
